+++ app/src/main/java/com/luckyzyx/luckytool/hook/scope/android/RemoveScrollToTopWhiteList.kt
@@ -11,15 +11,15 @@ object RemoveScrollToTopWhiteList : YukiBaseHooker() {
         val isEnable = prefs(ModulePrefs).getBoolean("remove_scroll_to_top_white_list", false)
         if (SDK < A13) return
         //Source OplusScrollToTopRusHelper -> OplusScrollToTopSystemManager
-        val helperClazz = "com.android.server.OplusScrollToTopRusHelper"
-        if (helperClazz.toClass().hasMethod { name = "isInWhiteList" }.not()) return
-        findClass(helperClazz).hook {
-            injectMember {
-                method {
-                    name = "isInWhiteList"
-                    paramCount = 1
+        findClass("com.android.server.OplusScrollToTopRusHelper").hook {
+            if (instanceClass.hasMethod { name = "isInWhiteList" }) {
+                injectMember {
+                    method {
+                        name = "isInWhiteList"
+                        paramCount = 1
+                    }
+                    if (isEnable) replaceToTrue()
                 }
-                if (isEnable) replaceToTrue()
             }
         }
     }