+++ app/src/main/java/com/luckyzyx/luckytool/hook/scope/systemui/VibrateWhenOpeningTheStatusBar.kt
@@ -2,8 +2,7 @@ package com.luckyzyx.luckytool.hook.scope.systemui
 
 import com.highcapable.yukihookapi.hook.entity.YukiBaseHooker
 import com.highcapable.yukihookapi.hook.factory.MembersType
-import com.luckyzyx.luckytool.utils.A13
-import com.luckyzyx.luckytool.utils.SDK
+import com.highcapable.yukihookapi.hook.factory.hasField
 
 object VibrateWhenOpeningTheStatusBar : YukiBaseHooker() {
     override fun onHook() {
@@ -11,30 +10,24 @@ object VibrateWhenOpeningTheStatusBar : YukiBaseHooker() {
         findClass("com.android.systemui.statusbar.phone.PanelViewController").hook {
             injectMember {
                 allMembers(MembersType.CONSTRUCTOR)
-                afterHook {
-                    field { name = "mVibrateOnOpening" }.get(instance).setTrue()
-                }
+                afterHook { field { name = "mVibrateOnOpening" }.get(instance).setTrue() }
             }
         }
         //Source PanelViewController -> config_vibrateOnIconAnimation
         findClass("com.android.systemui.statusbar.phone.StatusBar").hook {
             injectMember {
                 method { name = "start" }
-                afterHook {
-                    field { name = "mVibrateOnOpening" }.get(instance).setTrue()
-                }
+                afterHook { field { name = "mVibrateOnOpening" }.get(instance).setTrue() }
             }
         }
-        if (SDK < A13) return
         //Source StatusBarCommandQueueCallbacks -> config_vibrateOnIconAnimation
         findClass("com.android.systemui.statusbar.phone.StatusBarCommandQueueCallbacks").hook {
-            injectMember {
-                allMembers(MembersType.CONSTRUCTOR)
-                afterHook {
-                    field { name = "mVibrateOnOpening" }.get(instance).setTrue()
+            if (instanceClass.hasField { name = "mVibrateOnOpening" }) {
+                injectMember {
+                    allMembers(MembersType.CONSTRUCTOR)
+                    afterHook { field { name = "mVibrateOnOpening" }.get(instance).setTrue() }
                 }
             }
         }
-
     }
 }