+++ app/src/main/java/com/luckyzyx/luckytool/hook/hooker/HookSettings.kt
@@ -6,8 +6,8 @@ import com.luckyzyx.luckytool.hook.scope.settings.EnableStatusBarClockFormat
 import com.luckyzyx.luckytool.hook.scope.settings.ForceDisplayBottomGoogleSettings
 import com.luckyzyx.luckytool.hook.scope.settings.ForceDisplayDCBackLightMode
 import com.luckyzyx.luckytool.hook.scope.settings.HookAppDetails
-import com.luckyzyx.luckytool.hook.scope.settings.HookAppFeatureProvider
 import com.luckyzyx.luckytool.hook.scope.settings.HookIris5Controller
+import com.luckyzyx.luckytool.hook.scope.settings.HookSettingsFeature
 import com.luckyzyx.luckytool.hook.scope.settings.RemoveDpiRestartRecovery
 import com.luckyzyx.luckytool.hook.scope.settings.RemoveSettingsBottomLaboratory
 import com.luckyzyx.luckytool.hook.scope.settings.RemoveTopAccountDisplay
@@ -16,8 +16,8 @@ import com.luckyzyx.luckytool.utils.ModulePrefs
 
 object HookSettings : YukiBaseHooker() {
     override fun onHook() {
-        //HookAppFeatureProvider
-        loadHooker(HookAppFeatureProvider)
+        //HookSettingsFeature
+        loadHooker(HookSettingsFeature)
 
         //暗色模式列表
         if (prefs(ModulePrefs).getBoolean("dark_mode_list_enable", false)) {
@@ -57,9 +57,6 @@ object HookSettings : YukiBaseHooker() {
         //oplus_display_settings_screen_timeout_title
         //screen_timeout
         //dream_timeout_values
-
-        //HookSettingsFeature -> memc
-        //loadHooker(HookSettingsFeature)
 
         //persist.oplus.display.vrr
         //persist.oplus.display.vrr.adfr